fix #107 review comments
