+++ src/app/create-todo-form/create-todo-form.component.scss
@@ -5,13 +5,13 @@
   gap: 20px 0px;
   max-width: 500px;
 
-  input {
+  &__input {
     border: 1px solid #000;
     padding: 10px;
     border-radius: 10px;
   }
 
-  button {
+  &__btn {
     background-color: #F0BA4E;
     padding: 15px;
     border-radius: 10px;
@@ -21,5 +21,5 @@
 
 .alert {
   text-transform: uppercase;
-  color: red;
+  color: #FF0000;
 }
