+++ src/app/interfaces/todo-interface.ts
@@ -1,0 +1,6 @@
+export interface Todo {
+  userId: number | string;
+  id: number;
+  title: string;
+  completed: boolean;
+}
