+++ src/app/interfaces/user-interface.ts
@@ -1,0 +1,30 @@
+export interface User {
+  id: number;
+  name: string;
+  username?: string;
+  email: string;
+  address?: {
+    street: string;
+    suite: string;
+    city: string;
+    zipcode: string;
+    geo: {
+      lat: string;
+      lng: string;
+    };
+  };
+  phone?: string;
+  website: string;
+  company: {
+    name: string;
+    catchPhrase?: string;
+    bs?: string;
+  };
+}
+
+export interface CreateUser {
+  name: string;
+  email: string;
+  website: string;
+  companyName: string;
+}
