+++ src/app/todos-list/todos-list.component.scss
@@ -1,6 +1,6 @@
 .todos {
   display: flex;
   flex-wrap: wrap;
-  gap: 20px 0px;
+  gap: 20px 20px;
   margin: 20px 0;
 }
