+++ src/app/users-list/users-list.component.ts
@@ -4,37 +4,7 @@ import { UsersApiService } from '../users-api.service';
 import { UserCardComponent } from './user-card/user-card.component';
 import { UsersService } from '../users.service';
 import { CreateUserFormComponent } from '../create-user-form/create-user-form.component';
-
-export interface User {
-  id: number;
-  name: string;
-  username?: string;
-  email: string;
-  address?: {
-    street: string;
-    suite: string;
-    city: string;
-    zipcode: string;
-    geo: {
-      lat: string;
-      lng: string;
-    };
-  };
-  phone?: string;
-  website: string;
-  company: {
-    name: string;
-    catchPhrase?: string;
-    bs?: string;
-  };
-}
-
-interface UserFormData {
-  name: string;
-  email: string;
-  website: string;
-  companyName: string;
-}
+import { CreateUser } from '../interfaces/user-interface';
 
 @Component({
   selector: 'app-users-list',
@@ -58,7 +28,7 @@ export class UsersListComponent {
     this.usersService.deleteUser(id);
   }
 
-  createUser(formData: UserFormData) {
+  createUser(formData: CreateUser) {
     this.usersService.createUser({
       id: new Date().getDate(),
       name: formData.name,
