+++ src/app/users.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
-import { User } from './users-list/users-list.component';
+import { User } from './interfaces/user-interface';
 
 @Injectable({ providedIn: 'root' })
 export class UsersService {
